Properly test internal headers
Currently we werent testing all potential fallbacks we are using because some of the intrinsics are just available everywhere. This obviously led to some bugs slip in that would break everthing that includes type_traits.

To prevent this from happening again, add a script that crawls all internal headers and ensures both that they are building on their own and that all fallbacks path are at least exercised

This addresses nvbug4433362